add strict null checks, separate loader from ngo, rename utils
