+++ tsconfig.json
@@ -5,6 +5,7 @@
     "target": "es5",
     "outDir": "dist",
     "noEmitOnError": true,
+    "strictNullChecks":true,
     "noImplicitAny": true,
     "noUnusedParameters": true,
     "noUnusedLocals": true,
